+++ backend/microservices/user_management/user_management_service/__init__.py
@@ -1,4 +1,4 @@
-from flask import Flask, jsonify, request
+from flask import Flask, jsonify
 from flask_cors import CORS
 import os, json
 from dotenv import load_dotenv
@@ -47,14 +47,16 @@ def create_app(ENVIRONMENT='development'):
 
     app.register_blueprint(UserManagementService.main)
 
-    # TODO: Remove method before deploying
     @app.route('/db_reset', methods=['GET'])
     def db_create():
+        if ENVIRONMENT == 'production':
+            return jsonify({'message': 'Request denied.'}), 403
+
         db.session.execute(text("DROP TABLE IF EXISTS maga_user;"))
         db.session.commit()
         with app.app_context():
             db.create_all()
 
-        return 'Tables Reset!'
+        return jsonify({'message': 'Tables reset!'}), 200
 
     return app
